Eliminated duplication of test_adf_sequence()
