+++ vowpalwabbit/cb_adf.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <vector>
 
 LEARNER::base_learner* cb_adf_setup(VW::config::options_i& options, vw& all);
 
@@ -6,4 +7,5 @@ namespace CB_ADF
 {
 CB::cb_class get_observed_cost(multi_ex& examples);
 void global_print_newline(vw& all);
+example* test_adf_sequence(std::vector<example*>& ec_seq);
 }  // namespace CB_ADF
