+++ vowpalwabbit/cb_explore_adf.cc
@@ -93,36 +93,6 @@ void swap(T& ele1, T& ele2)
   ele1 = temp;
 }
 
-// Validates a multiline example collection as a valid sequence for action dependent features format.
-example* test_adf_sequence(multi_ex& ec_seq)
-{
-  if (ec_seq.size() == 0)
-    THROW("cb_adf: At least one action must be provided for an example to be valid.");
-
-  uint32_t count = 0;
-  example* ret = nullptr;
-  for (size_t k = 0; k < ec_seq.size(); k++)
-  {
-    example* ec = ec_seq[k];
-
-    // Check if there is more than one cost for this example.
-    if (ec->l.cb.costs.size() > 1)
-      THROW("cb_adf: badly formatted example, only one cost can be known.");
-
-    // Check whether the cost was initialized to a value.
-    if (ec->l.cb.costs.size() == 1 && ec->l.cb.costs[0].cost != FLT_MAX)
-    {
-      ret = ec;
-      count += 1;
-    }
-  }
-
-  if (count > 1)
-    THROW("cb_adf: badly formatted example, only one line can have a cost");
-
-  return ret;
-}
-
 // TODO: same as cs_active.cc, move to shared place
 float binary_search(float fhat, float delta, float sens, float tol = 1e-6)
 {
@@ -238,7 +208,7 @@ template <bool is_learn>
 void predict_or_learn_first(cb_explore_adf& data, multi_learner& base, multi_ex& examples)
 {
   // Explore tau times, then act according to optimal.
-  bool is_learn_example = is_learn && test_adf_sequence(examples) != nullptr;
+  bool is_learn_example = is_learn && CB_ADF::test_adf_sequence(examples) != nullptr;
   if (is_learn_example)
     multiline_learn_or_predict<true>(base, examples, data.offset);
   else
@@ -269,7 +239,7 @@ void predict_or_learn_greedy(cb_explore_adf& data, multi_learner& base, multi_ex
   data.offset = examples[0]->ft_offset;
   // Explore uniform random an epsilon fraction of the time.
 
-  if (is_learn && test_adf_sequence(examples) != nullptr)
+  if (is_learn && CB_ADF::test_adf_sequence(examples) != nullptr)
     multiline_learn_or_predict<true>(base, examples, data.offset);
   else
     multiline_learn_or_predict<false>(base, examples, data.offset);
@@ -292,7 +262,7 @@ void predict_or_learn_greedy(cb_explore_adf& data, multi_learner& base, multi_ex
 template <bool is_learn>
 void predict_or_learn_regcb(cb_explore_adf& data, multi_learner& base, multi_ex& examples)
 {
-  if (is_learn && test_adf_sequence(examples) != nullptr)
+  if (is_learn && CB_ADF::test_adf_sequence(examples) != nullptr)
   {
     for (size_t i = 0; i < examples.size() - 1; ++i)
     {
@@ -393,7 +363,7 @@ void predict_or_learn_bag(cb_explore_adf& data, multi_learner& base, multi_ex& e
   for (uint32_t i = 0; i < num_actions; i++) data.scores.push_back(0.f);
   vector<float>& top_actions = data.top_actions;
   top_actions.assign(num_actions, 0);
-  bool test_sequence = test_adf_sequence(examples) == nullptr;
+  bool test_sequence = CB_ADF::test_adf_sequence(examples) == nullptr;
   for (uint32_t i = 0; i < data.bag_size; i++)
   {
     // avoid updates to the random num generator
@@ -527,7 +497,7 @@ void predict_or_learn_cover(cb_explore_adf& data, multi_learner& base, multi_ex&
 template <bool is_learn>
 void predict_or_learn_softmax(cb_explore_adf& data, multi_learner& base, multi_ex& examples)
 {
-  if (is_learn && test_adf_sequence(examples) != nullptr)
+  if (is_learn && CB_ADF::test_adf_sequence(examples) != nullptr)
     multiline_learn_or_predict<true>(base, examples, data.offset);
   else
     multiline_learn_or_predict<false>(base, examples, data.offset);
@@ -639,7 +609,7 @@ void finish_multiline_example(vw& all, cb_explore_adf& data, multi_ex& ec_seq)
 template <bool is_learn>
 void do_actual_learning(cb_explore_adf& data, multi_learner& base, multi_ex& ec_seq)
 {
-  example* label_example = test_adf_sequence(ec_seq);
+  example* label_example = CB_ADF::test_adf_sequence(ec_seq);
   data.gen_cs.known_cost = CB_ADF::get_observed_cost(ec_seq);
 
   if (label_example == nullptr || !is_learn)
