+++ vowpalwabbit/explore_eval.cc
@@ -123,7 +123,7 @@ void finish_multiline_example(vw& all, explore_eval& data, multi_ex& ec_seq)
 template <bool is_learn>
 void do_actual_learning(explore_eval& data, multi_learner& base, multi_ex& ec_seq)
 {
-  example* label_example = CB_EXPLORE_ADF::test_adf_sequence(ec_seq);
+  example* label_example = CB_ADF::test_adf_sequence(ec_seq);
 
   if (label_example != nullptr)  // extract label
   {
